Allow passing scripts with -f or --from-file
When a script file is passed with -f or --from-file to :Jqplay, it will
be opened instead of a new jq scratch buffer.

The current implementation is a little bit strange and prone to failure.
We are writing to the same file that we are reading from: getbufline()
is passed to writefile()