Don't use any input buffer when -n option passed
When the -n or --null-input options are passed to jq, no input buffer is
read. The following commit adds support for this.

When :Jqplay is run with -n/--null-input, the jq filter and json output
buffer are always opened in a new tab page. The current json buffer is
ignored.

This partially solves #2.

TODO: We still need a global command that can be run from any buffer,
not just json.