Add support for --raw-input, make :Jqplay global
Make :Jqplay globally accessible so that it can be run from any filetype
not just json. This is useful when the -R/--raw-input option is passed.

The plugin will check if --null-input and --raw-input are used together.
When both options are passed to :Jqplay, the input buffer nevertheless
needs to be passed to jq despite the -n option. For an example, see
https://github.com/stedolan/jq/issues/1956