Make :Jqplay global command, not just buffer-local
Make :Jqplay globally accessible so that it can be run from any filetype
not just json. This is required when the -n/--null-input and/or
-R/--raw-input options are used.

While jq by default doesn't parse the input file when --null-input is
passed, there are situations when both, --null-input and an input file
are needed. One use case is when jq's "inputs" filter is used (see
https://github.com/stedolan/jq/issues/1956).

Another example when the input buffer is parsed by jq despite the
--null-input option is "seq 5 | jq -n '[inputs]'". Numbers can be passed
to jq directly without using the --raw-input option. The command will
output [1,2,4,5].

This example shows that it's not known in advance whether or not jq will
parse the passed input file when the --null-input option is used.

Therefore, whenever --null-input is passed to :Jqplay, the current
buffer will be *considered* as an input source for jq, i.e. the buffer
is passed to jq as stdin, even if jq won't parse it for some filters.

This approach has two disadvantages:
1. When the input buffer is deleted or wiped out, the jqplay session
   will be closed automatically. Imagine someone runs ":Jqplay -n" and
   doesn't care about the current buffer since s/he never intends to use
   the "inputs" filter. And during an open jqplay session decides to
   delete this buffer. This will unintentionally close the session.
2. Whenever the input buffer is modified jq will be automatically
   invoked.

See also discussion in https://github.com/bfrg/vim-jqplay/pull/3