Add -bang option to :JqplayScratch command
Running `:JqplayScratch!` with a bang will open an interactive session
in a new tab page and force the --null-input option, but without opening
an input buffer (internally `/dev/null` is passed to `jq` as stdin).
This means that `jq` filters like `inputs` won't work.

`:JqplayScratch!` and `:JqplayScratch! -n` are identical, the `-n` can
be omitted.

The current implementation is very messy and needs to be rewritten.